backend/requirements.txt
## The following requirements were added by pip freeze: Use ```pip install -r requirements```
absl-py==2.1.0
accelerate==0.27.2
annotated-types==0.6.0
asgiref==3.7.2
astunparse==1.6.3
base58==2.1.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cytoolz==0.12.3
dm-tree==0.1.8
ecdsa==0.18.0
eth-hash==0.7.0
eth-keys==0.5.0
eth-typing==4.0.0
eth-utils==2.3.1
filelock==3.13.1
Flask==3.0.2
flatbuffers==24.3.7
fsspec==2024.2.0
gast==0.5.4
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.1
grpcio-status==1.62.1
h5py==3.10.0
huggingface-hub==0.21.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
keras==3.0.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.2
scalecodec==1.2.8
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
sentencepiece==0.2.0
setuptools==69.1.1
six==1.16.0
substrate-interface==1.7.7
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.0rc0
tensorflow-intel==2.16.0rc0
termcolor==2.4.0
tf-keras==2.15.0
threadpoolctl==3.3.0
tokenizers==0.15.2
toolz==0.12.1
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
xxhash==3.4.1